requirements.txt
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==3.3.0
boto3==1.17.107
botocore==1.20.107
cassandra-driver==3.25.0
cffi==1.14.5
click==8.0.1
cycler==0.10.0
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
geomet==0.2.1.post1
greenlet==1.1.0
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
ipython-sql==0.4.0
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.20.3
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
prettytable==0.7.2
prometheus-client==0.11.0
prompt-toolkit==3.0.18
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.1.0
s3transfer==0.4.2
Send2Trash==1.5.0
six==1.16.0
SQLAlchemy==1.4.20
sqlparse==0.4.1
terminado==0.10.1
testpath==0.5.0
tornado==6.1
traitlets==5.0.5
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
